fix trilinear, and add a new multithreaded resampler.
the multithreaded one is broken, however... there is no ordering
between the threads, and yet the output writing assumes they'll
come out linearly.  ruh-roh.

also it really doesn't make any sense with the current interface,
since the interface forces both input and output volumes to be
in-core the whole time.