fix the multithreaded resampler.
no perf tests yet, not sure it's a great idea.  but at least it's
correct now, insofar as I can tell.

There were two issues.  The first was synchronization: because the
threads resampling can't close the channel (they can't know who is
the last thread), and the consumer thread can't close it---it's
ranging over the output channel, it's the thread that needs the
close in the first place!---that implies we need a third thread.
In practice I just put the consumer thread in a separate thread
instead of the main, and the main thread starts everything and
then waits for the consumers to finish.

The second issue was correctness/ordering: the scanlines get
output in arbitrary order, and yet the consumer was assuming they
arrived in-order.  Fixed by adding metadata to the scanline data
passed that identifies which scanline this is.

We copy the data in/out of the channel right now.  This seems less
than ideal, the parallel version does much more copying than the
serial version.  Hrum.